requirements.txt
numpy>=1.17
scipy
tqdm
lpips
mmcv
imageio
imageio-ffmpeg
opencv-python
torch_efficient_distloss
basicsr==1.3.5
addict
future
lmdb
Pillow
pyyaml
requests
scikit-image
tb-nightly
yapf
